chore(android): move onboarding/bowling pattern to shared designsystem module
