+++ android/core/designsystem/src/main/res/drawable/onboarding_pattern.xml
