+++ android/core/designsystem/src/main/res/values-night/colors.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+	<color name="bowling_pattern_background">#000000</color>
+</resources>
