--- android/feature/onboarding/ui/src/main/res/values/colors.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources xmlns:tools="http://schemas.android.com/tools" tools:ignore="UnusedResources">
-    <color name="container">#FFFFFFFF</color>
-</resources>
